central_management/server/requirements.txt
ocpp
websockets
python-dotenv
flask[async]
flask-cors
psycopg2-binary